+++ mcsc/constants.py
