+++ tests/data/1.15.2/expected_purge/DIM1/region/r.-1.-1.mca
