+++ tests/data/1.15.2/input/DIM-1/data/capabilities.dat
