+++ tests/data/1.15.2/input/DIM-1/poi/r.-1.-1.mca
