+++ tests/data/1.15.2/input/DIM1/data/capabilities.dat
