+++ tests/data/1.15.2/input/DIM1/poi/r.-1.0.mca
