+++ tests/data/1.15.2/input/data/capabilities.dat
