+++ tests/data/1.15.2/input/poi/r.-1.0.mca
