+++ tests/data/1.15.2/input/poi/r.0.-1.mca
