+++ tests/data/1.15.2/input/region/r.0.0.mca
